% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{agCensusInsurance}
\alias{agCensusInsurance}
\title{agCensusInsurance}
\format{
A data frame with 329783 rows and 10 columns covering Inf--Inf.
}
\source{
USDA NASS Quick Stats
}
\usage{
data(agCensusInsurance)
}
\description{
A combined dataset for agCensusInsurance
}
\keyword{datasets}
