% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copied_from_rmaADM.R
\name{clean_data}
\alias{clean_data}
\title{Apply standardized data cleaning opperations}
\source{
copied from https://github.com/dylan-turner25/rmaADM/blob/main/R/helpers.R
}
\usage{
clean_data(df)
}
\arguments{
\item{df}{a data frame to clean}
}
\value{
a cleaned data frame
}
\description{
Apply standardized data cleaning opperations
}
\examples{
\dontrun{clean_data(df)}
}
