% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm_cleaners.R
\name{clean_file_name2}
\alias{clean_file_name2}
\title{Clean the file names}
\usage{
clean_file_name2(file_names)
}
\arguments{
\item{file_names}{the file path of the file name to clean}
}
\value{
a version of the file name that is cleaned (i.e. snake case, .rds suffix, no extraneous information)
}
\description{
Clean the file names
}
\examples{
\dontrun{clean_file_name("2012_A01100_YieldAndTyield_YTD.txt")}
}
