% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{contiguous_county}
\alias{contiguous_county}
\title{Simulator Helper Datasets}
\format{
A data frame with 24307 rows and 12 columns covering Inf--Inf.
}
\source{
USDA-RMA, Actuarial Data Master - A0123
}
\usage{
data(contiguous_county)
}
\description{
A combined dataset for contiguous_county
}
\keyword{datasets}
