% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloaded_nass_large_datasets.R
\name{downloaded_nass_large_datasets}
\alias{downloaded_nass_large_datasets}
\title{Download and cache USDA NASS Quick Stats large dataset files}
\usage{
downloaded_nass_large_datasets(
  large_datasets,
  dir_dest = "./data-raw/fastscratch/nass/"
)
}
\arguments{
\item{large_datasets}{\verb{character list}
The base name of the Quick Stats file to download.  For example, use \code{"crops"} to fetch
\code{qs.crops_YYYYMMDD.txt.gz} or include \code{"census2022"} (e.g. \code{"census2022"}) to fetch the gzipped 2022 census version
(\code{qs.census2022.txt.gz}). any of:
"census2002","census2007","census2012","census2017","census2022",
"census2007zipcode","census2017zipcode",
"animals_products","crops","demographics","economics","environmental"}

\item{dir_dest}{\code{character(1)}
Path to a directory where downloaded files will be stored.  Defaults to \code{"./data-raw/fastscratch/nass/"}.}
}
\value{
Invisibly returns the normalized file large_dataset (e.g. \code{"qs.crops_YYYYMMDD.txt.gz"} or \code{"qs.censusYYYY.txt.gz"}) that was
downloaded or already present.
}
\description{
\code{downloaded_nass_large_datasets()} retrieves a Quick Stats file from the USDA National Agricultural Statistics Service (NASS)
https://www.nass.usda.gov/datasets/ page and saves it locally.  If the file is already present in the target directory, it is not re-downloaded.
}
\details{
\enumerate{
\item Prepends \code{"qs."} to the provided \code{large_dataset}.  If \code{large_dataset} contains \code{"census"}, appends \code{".txt.gz"},
otherwise \code{NULL}.
\item Ensures \code{dir_dest} exists (creates it if needed).
\item Scrapes the NASS datasets page (\verb{https://www.nass.usda.gov/datasets/}) for links ending in \code{.txt.gz}.
\item Downloads the matching file into \code{dir_dest} if not already present.
}
}
\examples{
\dontrun{
# Download the 'crops' dataset if not already cached:
downloaded_nass_large_datasets(large_dataset = "crops")

# Download the 2022 census version:
downloaded_nass_large_datasets(large_dataset = "census2022", 
dir_dest = "./data-raw/fastscratch/nass/")
}
}
\seealso{
Other USDA NASS Quick Stats: 
\code{\link{get_nass_census_data}()},
\code{\link{process_nass_dataset}()}
}
\concept{USDA NASS Quick Stats}
