% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_sobtpu_with_covariates.R
\name{enrich_with_commodity_grouping}
\alias{enrich_with_commodity_grouping}
\title{Add commodity groupings and TOPCROP tag (left join)}
\usage{
enrich_with_commodity_grouping(sob, commodity_map = NULL, top_n = 12L)
}
\arguments{
\item{sob}{data.frame/data.table with \code{commodity_year, commodity_code, insured_acres}}

\item{commodity_map}{Optional mapping with columns
\code{commodity_year, commodity_code, commodity_name, CROP, commodity_group}.
If \code{NULL}, it will be downloaded.}

\item{top_n}{integer. Number of crops to keep as named in \code{TOPCROP}; others -> "OTHER".}
}
\value{
data.table with \code{commodity_name}, \code{CROP}, \code{commodity_group}, \code{TOPCROP}
}
\description{
Add commodity groupings and TOPCROP tag (left join)
}
