% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_sobtpu_with_covariates.R
\name{enrich_with_election_grouping}
\alias{enrich_with_election_grouping}
\title{Add plan, product, coverage buckets, and unit structure labels}
\usage{
enrich_with_election_grouping(sob)
}
\arguments{
\item{sob}{data.frame/data.table with columns:
\itemize{
\item \code{insurance_plan_code}
\item \code{coverage_level_percent} (either 0-1 or 0-100)
\item \code{unit_structure_code}
}}
}
\value{
data.table with new columns: \code{PLAN}, \code{RPYP}, \code{COV}, \code{STRUCT}
}
\description{
Add plan, product, coverage buckets, and unit structure labels
}
