% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_sobtpu_with_covariates.R
\name{enrich_with_location_grouping}
\alias{enrich_with_location_grouping}
\title{Add state names/abbreviations, ERS Region, and NASS CRD (left joins)}
\usage{
enrich_with_location_grouping(sob, ers_xls = NULL, crd_csv = NULL)
}
\arguments{
\item{sob}{data.frame/data.table with \code{state_code, county_code}}

\item{ers_xls}{Optional path to the ERS excel file; if \code{NULL}, it is downloaded.}

\item{crd_csv}{Optional path to the CRD csv file; if \code{NULL}, it is downloaded.}
}
\value{
data.table with \code{state_abbreviation}, \code{state_name}, \code{ERSReg}, \code{ERSReg_cd}, \code{CRD}
}
\description{
Add state names/abbreviations, ERS Region, and NASS CRD (left joins)
}
