% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_sobtpu_with_covariates.R
\name{enrich_with_practice_grouping}
\alias{enrich_with_practice_grouping}
\title{Add irrigation/organic practice recodes (left join)}
\usage{
enrich_with_practice_grouping(sob, practice_map = NULL)
}
\arguments{
\item{sob}{data.frame/data.table with \code{commodity_year, commodity_code, practice_code}}

\item{practice_map}{Optional data.frame/data.table with columns
\code{commodity_year, commodity_code, practice_code, irrigation_recode, organic_recode}.
If \code{NULL}, it will be downloaded from the project release.}
}
\value{
data.table with \code{irrigation_recode}, \code{organic_recode}
}
\description{
Add irrigation/organic practice recodes (left join)
}
