% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_sobtpu_with_covariates.R
\name{enrich_with_structural_breaks}
\alias{enrich_with_structural_breaks}
\title{Add farm-bill era labels and a before/after 2012 combo}
\usage{
enrich_with_structural_breaks(sob)
}
\arguments{
\item{sob}{data.frame/data.table with \code{commodity_year}}
}
\value{
data.table with \code{period_farmbill}, \code{period_combo}
}
\description{
Add farm-bill era labels and a before/after 2012 combo
}
