% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{fcip_aph_base_rate}
\alias{fcip_aph_base_rate}
\title{fcip_aph_base_rate}
\format{
A data frame with 428502 rows and 6 columns covering 2001-2026.
}
\source{
USDA-RMA, Actuarial Data Master supplemented data from legacy ADM files
}
\usage{
data(fcip_aph_base_rate)
}
\description{
A combined dataset for fcip_aph_base_rate
}
\keyword{datasets}
