% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{FCIP_FORCE_CHARACTER_KEYS}
\alias{FCIP_FORCE_CHARACTER_KEYS}
\title{Column names to coerce to character}
\format{
A \code{character} vector of column names.
}
\usage{
FCIP_FORCE_CHARACTER_KEYS
}
\value{
A \code{character} vector of field names to coerce to character
}
\description{
A character vector of column names that should be converted to character
during data ingestion and cleaning.
}
\keyword{datasets}
