% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{FCIP_FORCE_NUMERIC_KEYS}
\alias{FCIP_FORCE_NUMERIC_KEYS}
\title{Column names to coerce to numeric}
\format{
A \code{character} vector of column names.
}
\usage{
FCIP_FORCE_NUMERIC_KEYS
}
\value{
A \code{character} vector of field names to coerce to numeric.
}
\description{
A character vector of column names that should be converted from character
to numeric during data ingestion and cleaning.
}
\details{
The following fields, although often stored as text, represent numeric
values and must be coerced for accurate calculation and analysis:
\itemize{
\item \code{commodity_year}: Crop year of the record.
\item Fields in \code{FCIP_INSURANCE_POOL}:
\code{state_code}, \code{county_code}, \code{commodity_code},
\code{type_code}, \code{practice_code}.
\item \code{record_category_code}: Category of the record.
\item \code{insurance_plan_code}: Code for the insurance plan.
\item \code{coverage_level_percent}: Coverage level percentage.
}
}
\examples{
\dontrun{
# View default keys
FCIP_FORCE_NUMERIC_KEYS

# Extend with a custom numeric field
rFarmPolicySim:::FCIP_FORCE_NUMERIC_KEYS <- c(
  rFarmPolicySim:::FCIP_FORCE_NUMERIC_KEYS,
  "custom_numeric_field"
)
}

}
\keyword{datasets}
