% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_fcip.R
\name{get_yu2018_instrument}
\alias{get_yu2018_instrument}
\title{Formulate & Merge National Subsidy Rate Instrument (Yu et al., 2018)}
\usage{
get_yu2018_instrument(
  dt,
  delivery_systems = c("RBUP", "FBUP"),
  plan_codes = c(1:3, 90, 44, 25, 42),
  coverage_levels = c(65, 75)
)
}
\arguments{
\item{dt}{sobcov}

\item{delivery_systems}{Character vector. Delivery systems to include;
default \code{c("RBUP","FBUP")}.}

\item{plan_codes}{Integer vector. Insurance plan codes to include;
default \code{c(1:3, 90, 44, 25, 42)}.}

\item{coverage_levels}{Numeric vector. Percent coverage levels to keep;
default \code{c(65, 75)}.}
}
\value{
A \link[data.table]{data.table} in wide form with:
\item{crop_yr}{Crop year.}
\item{subsidy_rate_65}{National subsidy rate at 65\\% coverage (if requested).}
\item{subsidy_rate_75}{National subsidy rate at 75\\% coverage (if requested).}
}
\description{
Downloads the historical “Summary of Business” RDS and computes
national subsidy‐rate instruments at specified coverage levels,
following Yu et al. (2018).
}
