% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fcip.R
\name{harmonize_codes_and_names}
\alias{harmonize_codes_and_names}
\title{Harmonize names/codes in a data table of insurance elections}
\usage{
harmonize_codes_and_names(df)
}
\arguments{
\item{df}{A \code{data.frame}}
}
\value{
A \code{data.frame} with the same columns as \code{df} with new columns
}
\description{
This function looks for two columns and creates recoded versions grouping similar
codes into broader categories.
}
