% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_fcip.R
\name{harmonize_election_codes}
\alias{harmonize_election_codes}
\title{Harmonize election codes in a data table of insurance elections}
\usage{
harmonize_election_codes(df)
}
\arguments{
\item{df}{A \code{data.frame} containing, at minimum,
one or both of the columns:
\itemize{
\item \code{unit_structure_code}
\item \code{insurance_plan_code}
}}
}
\value{
A \code{data.frame} with the same columns as \code{df} with new columns:
\itemize{
\item \code{unit_structure_recode} (if \code{unit_structure_code} was present)
\item \code{insurance_plan_recode} (if \code{insurance_plan_code} was present)
}
}
\description{
This function looks for two columns—\code{unit_structure_code} and
\code{insurance_plan_code}—and creates recoded versions grouping similar
codes into broader categories.
}
