% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{ice_administrative_fee_waiver_code}
\alias{ice_administrative_fee_waiver_code}
\title{ice_administrative_fee_waiver_code}
\format{
A data frame with 66 rows and 4 columns covering Inf--Inf.
}
\source{
USDA-RMA, Insurance Control Elements - PASS - D00154
}
\usage{
data(ice_administrative_fee_waiver_code)
}
\description{
A combined dataset for ice_administrative_fee_waiver_code
}
\keyword{datasets}
