% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{ice_ao_expense_subsidy_percent}
\alias{ice_ao_expense_subsidy_percent}
\title{ice_ao_expense_subsidy_percent}
\format{
A data frame with 4232 rows and 7 columns covering 1998-2025.
}
\source{
USDA-RMA, Insurance Control Elements - PASS - D00154
}
\usage{
data(ice_ao_expense_subsidy_percent)
}
\description{
A combined dataset for ice_ao_expense_subsidy_percent
}
\keyword{datasets}
