% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{ice_policy_history_request_code}
\alias{ice_policy_history_request_code}
\title{ice_policy_history_request_code}
\format{
A data frame with 127 rows and 4 columns covering Inf--Inf.
}
\source{
USDA-RMA, Insurance Control Elements - PASS - D00154
}
\usage{
data(ice_policy_history_request_code)
}
\description{
A combined dataset for ice_policy_history_request_code
}
\keyword{datasets}
