% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{ice_program_indicator}
\alias{ice_program_indicator}
\title{ice_program_indicator}
\format{
A data frame with 46 rows and 4 columns covering Inf--Inf.
}
\source{
USDA-RMA, Insurance Control Elements - PASS - D00154
}
\usage{
data(ice_program_indicator)
}
\description{
A combined dataset for ice_program_indicator
}
\keyword{datasets}
