% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{ice_yield_type_code}
\alias{ice_yield_type_code}
\title{ice_yield_type_code}
\format{
A data frame with 913 rows and 15 columns covering Inf--Inf.
}
\source{
USDA-RMA, Insurance Control Elements - PASS - D00154
}
\usage{
data(ice_yield_type_code)
}
\description{
A combined dataset for ice_yield_type_code
}
\keyword{datasets}
