% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{marketing_year_avg_price}
\alias{marketing_year_avg_price}
\title{marketing_year_avg_price}
\format{
A data frame with 24396 rows and 5 columns covering 1866-2024.
}
\source{
USDA NASS Quick Stats
}
\usage{
data(marketing_year_avg_price)
}
\description{
A combined dataset for marketing_year_avg_price
}
\keyword{datasets}
