% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{nassSurveyAnimalInventory}
\alias{nassSurveyAnimalInventory}
\title{nassSurveyAnimalInventory}
\format{
A data frame with 175547 rows and 10 columns covering 1920-2025.
}
\source{
USDA NASS Quick Stats - Animal inventory as of first of Jan
}
\usage{
data(nassSurveyAnimalInventory)
}
\description{
A combined dataset for nassSurveyAnimalInventory
}
\keyword{datasets}
