% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{rma_type_recodes}
\alias{rma_type_recodes}
\title{rma_type_recodes}
\format{
A data frame with 230502 rows and 7 columns covering 1999-2025.
}
\source{
USDA-RMA, Summary of business by type, practice, and unit structure
}
\usage{
data(rma_type_recodes)
}
\description{
A combined dataset for rma_type_recodes
}
\keyword{datasets}
