% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{update_packages}
\alias{update_packages}
\title{Install a GitHub R Package Only If Its R/ Folder Has Newer Changes}
\usage{
update_packages(user, repo, package)
}
\arguments{
\item{user}{GitHub username or organization (string), e.g. \code{"dylan-turner25"}.}

\item{repo}{Repository name (string), e.g. \code{"rmaADM"}.}

\item{package}{Repository name (string), e.g. \code{"rmaADM"}.}
}
\value{
Invisibly returns \code{TRUE} if installation was performed, \code{FALSE} otherwise.
}
\description{
Compares the most recent commit date in the \code{R/} directory of a GitHub repo
against the modification times of your locally installed packages \code{R/*.R} files.
If the GitHub version is newer (or the package is not installed), installs via
\code{remotes::install_github()}.
}
\details{
\itemize{
\item Requires the \pkg{gh} package to query GitHub commits API and the \pkg{remotes}
package to install from GitHub. Installs them if missing.
\item If the package \code{repo} is not already installed locally, this function treats
it as, very out of date, and will install unconditionally.
}
}
\examples{
\dontrun{
# Check and update rmaADM from dylan-turner25 if needed:
update_packages(user="dylan-turner25", repo="rmaADM",package="rmaADM")
update_packages(user="dylan-turner25", repo="rfcip",package="rmaADM")
update_packages(user="ftsiboe", repo="US-FarmSafetyNet-Lab",package="USFarmSafetyNetLab")
}

}
