% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USFarmSafetyNetLab-package.R
\docType{package}
\name{USFarmSafetyNetLab-package}
\alias{USFarmSafetyNetLab}
\alias{USFarmSafetyNetLab-package}
\title{USFarmSafetyNetLab: US Farm Safety Net Lab}
\description{
| This repository centralizes research outputs, analytical tools, and resources exploring and evaluating the United States agricultural safety net programs Supported programs include: * Federal Crop Insurance Program (FCIP) * Noninsured Crop Disaster Assistance Program (NAP) * Price Loss Coverage (PLC) * Agricultural Risk Coverage (ARC) * Ad-hoc disaster assistance programs Users can run Monte Carlo experiments to evaluate program payments, subsidies, and election decisions under alternative policy scenarios.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/you/US-FarmSafetyNet-Lab}
  \item Report bugs at \url{https://github.com/you/US-FarmSafetyNet-Lab/issues}
}

}
\author{
\strong{Maintainer}: Francis Tsiboe \email{ftsiboe@hotmail.com} (\href{https://orcid.org/0000-0001-5984-1072}{ORCID})

}
\keyword{internal}
