% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_data.R
\docType{data}
\name{state_rental_rates}
\alias{state_rental_rates}
\title{state_rental_rates}
\format{
A data frame with 1736 rows and 5 columns covering 1994-2024.
}
\source{
Output from get_state_rental_rates() function
}
\usage{
data(state_rental_rates)
}
\description{
A combined dataset for state_rental_rates
}
\keyword{datasets}
