% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USFarmSafetyNetLab-package.R
\docType{package}
\name{USFarmSafetyNetLab-package}
\alias{USFarmSafetyNetLab}
\alias{USFarmSafetyNetLab-package}
\title{USFarmSafetyNetLab: US Farm Safety Net Lab}
\description{
This repository centralizes research outputs, analytical tools, and resources for exploring and evaluating the United States agricultural safety net programs. It supports analysis of key programs including the Federal Crop Insurance Program (FCIP), the Noninsured Crop Disaster Assistance Program (NAP), Price Loss Coverage (PLC), Agricultural Risk Coverage (ARC), and various ad-hoc disaster assistance programs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/you/USFarmSafetyNetLab}
  \item Report bugs at \url{https://github.com/you/USFarmSafetyNetLab/issues}
}

}
\keyword{internal}
